fixes all missing tests
