+++ Source/Constants.swift
@@ -7,13 +7,5 @@
 
 import Foundation
 
-let CONFLICTS = "_am_conflicts_"
-let OBJECT_ID = "_am_objectId_"
-let LIST_VALUES = "_am_list_values_"
-let COUNTER_VALUE = "_am_counter_value_"
-let TABLE_VALUES = "_am_tabel_values_"
-let CACHE = "_am_cache_"
-let ISTEXT = "_am_is_Text_Element_"
-
 let ROOT_ID: String = "00000000-0000-0000-0000-000000000000"
